Move "getGridSizeForLevel" from grid to utils
